Bilateral swap: Implemented the ABI input encoder for a zk bilateral swap + directory pattern reorganisation + inputOwner inclusion in the encoding
Reorganised directory to make use of the index.js pattern and rearranged the files associated with particular proofs into their own folder

Included the AZTEC inputOwners in the ABI input encoding, despite this not being required in the equivalent joinSplit ABI file. This was done because in a joinSplit tx, the input signatures are provided. This allows the input AZTEC note owner addresses to be reconstructed from ECDSA signature params. However, for a bilateral swap; input ECDSA signatures are not provided. This makes it necessary to explicit provide the inputOwner addresses in the input ABI encoding