Implemented PR feedback for BilateralSwap.sol.
Changes mainly revolve around hard coding in particular memory addresses. This is possible in a Bilateral Swap, but not in a join split, because we know in advance the total number of notes involved (n = 4). So we don't have arrays of undetermined length.

Hard coding in memory addresses saves on some gas