Implement PR feedback on duplicate outputCoder.js script, in abiEncoder of aztec.js
Earlier merging of the branches of various proofs resulted in an outputCoder.js script and associated test for each proof.
This is unnecessary - the script performs the exact same functionality.

Removed these duplicates, moved the script to the top of the local directory and exported.

Updated affected import filepaths in other parts of the codebase