Implement PR feedback - encoderFactory; new way of organising the ABI encoding
There is now a template for the ABI encoding, that should be adaptable to different zero knowledge proofs. This significantly decreases the amount of duplicated code - proofs will use this with little moderation

Implemented it for the bilateralSwap, tests pass and updated import file paths. Will be removing the various subfolders for the different proofs in the abiEncoder once all implemented.

Removed redundant tests

Also, seperated HexString class out into seperate module - will be moving this into a better location in a later tidying up commit