Change the error handling approach in the proof verification script. Changing from immediate throwing of errors using customError to the error array approach.
This is different to my previous approach in this PR. Realised that immediately explicitly throwing the error in verification makes it much harder to thoroughly test that the verification algo works as expected. For example, with the random proof data test, the aim is to check that a CHALLENGE_RESPONSE_FAIL error is thrown. However, before getting to this error there will likely be a multitude of other errors likely thrown - e.g.  too big a scalar value. If we stick to the throw explicit error immediately model, it's impossible to check that the proof failed for the correct reason.

However, I am sticking with the immediate throwing of errors in the proof construction code - here it is important that the transaction throws as soon as there is a problem and the user can then act on this feedback

Moved hexToGroupScalar() in joinSplit/verifier.js to proofUtils.js - I will be refactoring the bilateralSwap and dividendComputation proofs to use this method. It allows the form of the challenge to be checked.

Some comment removal